Disable recursive route check for GCE compatibility.
GCE uses a /32 for VM IPs with a default gateway that is, thus,
off-subnet.  This patch removes a check in BIRD that prevents
BIRD from accepting such a route as a valid next hop.

The following trail asserts that the check is not really needed
but that it is normally a useful sanity check:

https://www.mail-archive.com/bird-users@atrey.karlin.mff.cuni.cz/msg02600.html

Ideally GCE would set the online flag on the route or use a
link-local IP instead but that is unlikely to change.