Fix change detection logic for tunnel routes.
When a tunnel route gets added to the kernel, the gateway gets replaced
with the source IP, and the interface gets replaced with the tunnel
interface. The change detection wasn't taking this into account, causing
krt_same_dest return false, and the route would get deleted and
readded constantly.

This commit fixes it by not doing the krt_same_dest check for tunnel
routes, doing its own checks instead.