Build BIRD container for dual ToR and service advertisement tests
Here we build a BIRD container image that is useful for dual ToR and
service advertisement testing.

- For dual ToR the requirement is for a BIRD that models the ToRs and
any other infrastructure routers.  This is primarily IPv4 only,
because we have not yet looked at IPv6 dual ToR.

- For service advertisement the requirement is for a BIRD that models
an external router that should receive the service IP routes that
Calico advertises.  This is for both IPv4 and IPv6.