[WiP] support cross-compiling for 32-bit arm v7
This is a little cumbersome since the host triplet is distro-specific.
On Debian e.g. Dockerfile-cross - the compilers are named by the Debian
host triplet - arm-linux-gnueabihf for armv7!

The existing compiler name generation in create_binaries.sh is not
currently compatible. A nice clean-up would be to always use the host
triplet, no matter if somethign is a native or cross build - on Debian
the system compiler will always be available by its long name.

Signed-off-by: Josua Mayer <josua@solid-run.com>