Provide tweaks and early BGP function needed for dual ToR operation
This is a proof of concept where a modified calico/bird container
image is used to perform start-of-day tweaks needed for dual ToR, and
to provide "early BGP" function, to export the local node's stable
address, before calico-node's own BGP starts running.

Please note: this is not - at all - necessarily the final product
architecture for doing this.

Some details of note:

- Allow getting router ID from more interface names than just eth0, as
  they are named "enp0s*" in my test rig.

- Early BIRD listens on port 8179 so as not to clash with calico-node.

- Currently assumes a particular IP numbering scheme, which will need
  generalizing before anything like this is productized.